refactor(database): update cargo_capacity type and add departureRoutes relation
- Introduce departureRoutes relation in Airport model for easier access to related routes.

+++ database/migrations/2026_05_12_165029_create_basic_tables.php
@@ -44,7 +44,7 @@ return new class extends Migration
             $table->char('icao', 4)->unique();
             $table->unsignedSmallInteger('range_nm')->nullable();
             $table->unsignedSmallInteger('pax_capacity')->nullable();
-            $table->unsignedInteger('cargo_capacity')->nullable();
+            $table->unsignedSmallInteger('cargo_capacity')->nullable();
             $table->string('image_url')->nullable();
             $table->timestamps();
         });
